feat(newdaysafety): major auth changes and bug fixes
