+++ .env
@@ -11,3 +11,6 @@ ADMIN_EMAIL="newdaysafety@gmail.com"
 EMAIL_PASS="qtam bcjs rncf puwi"
 NEXTAUTH_SECRET="rajeshTheGreat"
 
+GOOGLE_CLIENT_ID=970089284011-ckiai7i2igaonahio40mkqm3pcqckdjc.apps.googleusercontent.com
+GOOGLE_CLIENT_SECRET=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
+NEXTAUTH_URL=http://localhost:3000
