+++ app/api/order/route.ts
@@ -10,14 +10,18 @@ export const dynamic = "auto";
 export async function POST(request: NextRequest) {
   try {
     const body = await request.json();
-    const authResult = await checkRole([UserRole.ADMIN, UserRole.USER])(
+    const authResult = await checkRole(
+      [UserRole.ADMIN, UserRole.USER],
       request
     );
 
     if (authResult instanceof Response) {
       return authResult;
     }
-    const response = await controller.createOrder(authResult.User.UserId, body);
+    const response = await controller.createOrder(
+      authResult.User.UserId!,
+      body
+    );
     return response;
   } catch (error) {
     console.error("Route Error:", error);
@@ -34,7 +38,8 @@ export async function POST(request: NextRequest) {
 
 export async function GET(request: NextRequest) {
   try {
-    const authResult = await checkRole([UserRole.ADMIN, UserRole.USER])(
+    const authResult = await checkRole(
+      [UserRole.ADMIN, UserRole.USER],
       request
     );
 
@@ -42,8 +47,8 @@ export async function GET(request: NextRequest) {
       return authResult;
     }
     const response = await controller.getOrderDetails(
-      authResult.User.UserId,
-      authResult.User.Role
+      authResult.User.UserId!,
+      authResult.User.Role!
     );
     return response;
   } catch (error) {
@@ -62,7 +67,8 @@ export async function GET(request: NextRequest) {
 export async function PUT(request: NextRequest) {
   try {
     const body = await request.json();
-    const authResult = await checkRole([UserRole.ADMIN, UserRole.USER])(
+    const authResult = await checkRole(
+      [UserRole.ADMIN, UserRole.USER],
       request
     );
 
@@ -70,7 +76,7 @@ export async function PUT(request: NextRequest) {
       return authResult;
     }
     const response = await controller.updateOrderstatus(
-      authResult.User.UserId,
+      authResult.User.UserId!,
       body
     );
     return response;
