+++ app/api/user/route.ts
@@ -10,7 +10,8 @@ export const dynamic = "auto";
 export async function POST(request: NextRequest) {
   try {
     const body = await request.json();
-    const authResult = await checkRole([UserRole.ADMIN, UserRole.USER])(
+    const authResult = await checkRole(
+      [UserRole.ADMIN, UserRole.USER],
       request
     );
 
@@ -18,7 +19,7 @@ export async function POST(request: NextRequest) {
       return authResult;
     }
     const response = await controller.updateProfile(
-      authResult.User.UserId,
+      authResult.User.UserId!,
       body
     );
     return response;
@@ -37,14 +38,15 @@ export async function POST(request: NextRequest) {
 
 export async function GET(request: NextRequest) {
   try {
-    const authResult = await checkRole([UserRole.ADMIN, UserRole.USER])(
+    const authResult = await checkRole(
+      [UserRole.ADMIN, UserRole.USER],
       request
     );
 
     if (authResult instanceof Response) {
       return authResult;
     }
-    const response = await controller.getProfile(authResult.User.UserId);
+    const response = await controller.getProfile(authResult.User.UserId!!);
     return response;
   } catch (error) {
     console.error("Route Error:", error);
