+++ app/auth-old/signup/page.tsx
@@ -1,0 +1,148 @@
+"use client";
+
+import { useForm } from "react-hook-form";
+import { zodResolver } from "@hookform/resolvers/zod";
+import { useMutation } from "@tanstack/react-query";
+import { signUp } from "@/lib/FE/api";
+import { SignUpInput, signUpSchema } from "@/lib/FE/validations/auth";
+import { toast } from "sonner";
+import Link from "next/link";
+import { useRouter } from "next/navigation";
+import { UserPlus } from "lucide-react";
+
+export default function SignUpPage() {
+  const router = useRouter();
+  const {
+    register,
+    handleSubmit,
+    formState: { errors },
+  } = useForm<SignUpInput>({
+    resolver: zodResolver(signUpSchema),
+  });
+
+  const mutation = useMutation({
+    mutationKey: ["signUp"],
+    mutationFn: signUp,
+    onSuccess: (data) => {
+      if (data.error) {
+        toast.error(`something went wrong ${data.message}`);
+      } else {
+        toast.success("Account created successfully!");
+        router.push("/auth/signin");
+      }
+    },
+    onError: (error) => {
+      toast.error("Failed to create account. Please try again.");
+    },
+  });
+
+  const onSubmit = (data: SignUpInput) => {
+    mutation.mutate(data);
+  };
+
+  return (
+    <div className="min-h-screen bg-gradient-to-br from-gray-900 to-gray-800 flex items-center justify-center p-4">
+      <div className="max-w-md w-full bg-white rounded-xl shadow-xl p-8">
+        <div className="flex flex-col items-center mb-8">
+          <h1 className="text-2xl font-bold text-gray-900">Create Account</h1>
+          <p className="text-gray-600 text-sm">Join us today!</p>
+        </div>
+
+        <form onSubmit={handleSubmit(onSubmit)} className="space-y-4">
+          <div className="grid grid-cols-2 gap-4">
+            <div>
+              <label className="block text-sm font-medium text-gray-700">
+                First Name
+              </label>
+              <input
+                {...register("FirstName")}
+                className="mt-1 block w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm focus:outline-none focus:ring-blue-500 focus:border-blue-500"
+              />
+              {errors.FirstName && (
+                <p className="mt-1 text-xs text-red-600">
+                  {errors.FirstName.message}
+                </p>
+              )}
+            </div>
+            <div>
+              <label className="block text-sm font-medium text-gray-700">
+                Last Name
+              </label>
+              <input
+                {...register("LastName")}
+                className="mt-1 block w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm focus:outline-none focus:ring-blue-500 focus:border-blue-500"
+              />
+              {errors.LastName && (
+                <p className="mt-1 text-xs text-red-600">
+                  {errors.LastName.message}
+                </p>
+              )}
+            </div>
+          </div>
+          <div>
+            <label className="block text-sm font-medium text-gray-700">
+              Email
+            </label>
+            <input
+              {...register("Email")}
+              type="email"
+              className="mt-1 block w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm focus:outline-none focus:ring-blue-500 focus:border-blue-500"
+            />
+            {errors.Email && (
+              <p className="mt-1 text-xs text-red-600">
+                {errors.Email.message}
+              </p>
+            )}
+          </div>
+          <div>
+            <label className="block text-sm font-medium text-gray-700">
+              Username
+            </label>
+            <input
+              {...register("Username")}
+              className="mt-1 block w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm focus:outline-none focus:ring-blue-500 focus:border-blue-500"
+            />
+            {errors.Username && (
+              <p className="mt-1 text-xs text-red-600">
+                {errors.Username.message}
+              </p>
+            )}
+          </div>
+          <div>
+            <label className="block text-sm font-medium text-gray-700">
+              Password
+            </label>
+            <input
+              {...register("Password")}
+              type="password"
+              className="mt-1 block w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm focus:outline-none focus:ring-blue-500 focus:border-blue-500"
+            />
+            {errors.Password && (
+              <p className="mt-1 text-xs text-red-600">
+                {errors.Password.message}
+              </p>
+            )}
+          </div>
+
+          <button
+            type="submit"
+            disabled={mutation.isPending}
+            className="w-full py-2 px-4 border border-transparent rounded-md shadow-sm text-sm font-medium text-white bg-blue-600 hover:bg-blue-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-blue-500 disabled:opacity-50 disabled:cursor-not-allowed"
+          >
+            {mutation.isPending ? "Creating account..." : "Sign Up"}
+          </button>
+        </form>
+
+        <p className="mt-4 text-center text-sm text-gray-600">
+          Already have an account?{" "}
+          <Link
+            href="/auth/signin"
+            className="text-blue-600 hover:text-blue-500"
+          >
+            Sign in
+          </Link>
+        </p>
+      </div>
+    </div>
+  );
+}
