+++ app/auth/signin/components/SignInForm.tsx
@@ -15,7 +15,6 @@ export function SignInForm() {
     const formData = new FormData(e.currentTarget);
     const email = formData.get("email") as string;
     const password = formData.get("password") as string;
-    console.log(email, password);
 
     try {
       const result = await signIn("credentials", {
