--- app/auth1/signin/page.tsx
@@ -1,17 +1,0 @@
-"use client";
-import { SignInForm } from "./components/SignInForm";
-
-export default function SignInPage() {
-  return (
-    <div className="min-h-screen flex items-center justify-center bg-gray-50 py-12 px-4 sm:px-6 lg:px-8">
-      <div className="max-w-md w-full space-y-8">
-        <div>
-          <h2 className="mt-6 text-center text-3xl font-extrabold text-gray-900">
-            Sign in to your account
-          </h2>
-        </div>
-        <SignInForm />
-      </div>
-    </div>
-  );
-}
