+++ components/molecules/ProductForm/ColorVariant/imageUploader.tsx
@@ -18,10 +18,8 @@ export function ImageUploader({
     e: React.ChangeEvent<HTMLInputElement>
   ) => {
     const files = e.target.files;
-    console.log(files);
     if (files) {
       const imageUrls = await handleImageUpload(Array.from(files));
-      console.log(imageUrls);
       onImagesChange([...images, ...imageUrls]);
     }
   };
