+++ components/molecules/cart/CartItemCard.tsx
@@ -18,7 +18,6 @@ export function CartItemCard({
   onUpdateQuantity,
   onRemove,
 }: CartItemCardProps) {
-  console.log(item);
   return (
     <motion.div
       initial={{ opacity: 0, y: 20 }}
