+++ lib/prisma.ts
@@ -1,5 +1,20 @@
-// lib/prisma.ts
+// // lib/prisma.ts
+// import { PrismaClient } from "@prisma/client";
+
+// const prisma = new PrismaClient();
+// export default prisma;
 import { PrismaClient } from "@prisma/client";
 
-const prisma = new PrismaClient();
+const prismaClientSingleton = () => {
+  return new PrismaClient();
+};
+
+declare global {
+  var prisma: undefined | ReturnType<typeof prismaClientSingleton>;
+}
+
+const prisma = globalThis.prisma ?? prismaClientSingleton();
+
+if (process.env.NODE_ENV !== "production") globalThis.prisma = prisma;
+
 export default prisma;
